+++ lib/fly_chat_platform_interface.dart
@@ -233,7 +233,7 @@ abstract class FlyChatFlutterPlatform extends PlatformInterface {
     throw UnimplementedError('has not been implemented.');
   }
 
-  Future<dynamic> handleReceivedMessage(Map notificationdata) {
+  Future<dynamic> handleReceivedMessage(Map notificationData) {
     throw UnimplementedError('has not been implemented.');
   }
 
@@ -339,7 +339,7 @@ abstract class FlyChatFlutterPlatform extends PlatformInterface {
     throw UnimplementedError('has not been implemented.');
   }
 
-  Future<dynamic> registerUser(String userIdentifier, {String token = ""}) {
+  Future<dynamic> registerUser(String userIdentifier, {String fcmToken = ""}) {
     throw UnimplementedError('has not been implemented.');
   }
 
@@ -586,7 +586,8 @@ abstract class FlyChatFlutterPlatform extends PlatformInterface {
     String? messageId,
     double? messageTime,
     bool? exclude,
-    int limit = 25}) {
+    int limit = 25,
+    bool ascendingOrder = true}) {
     throw UnimplementedError('initializeMessageList has not been implemented.');
   }
 
@@ -636,7 +637,7 @@ abstract class FlyChatFlutterPlatform extends PlatformInterface {
     throw UnimplementedError('has not been implemented.');
   }
 
-  Future<bool?> refreshAndGetAuthToken() {
+  Future<String?> refreshAndGetAuthToken() {
     throw UnimplementedError('has not been implemented.');
   }
 
@@ -1024,6 +1025,9 @@ abstract class FlyChatFlutterPlatform extends PlatformInterface {
   }
   Future<bool?> isUserVideoMuted([String? userJid]) async {
     throw UnimplementedError('isUserVideoMuted has not been implemented.');
+  }
+  Future<String?> openAudioFilePicker() async {
+    throw UnimplementedError('openAudioFilePicker has not been implemented.');
   }
 
 }