+++ lib/flychat.dart
@@ -22,7 +22,7 @@ class Mirrorfly {
       String? storageFolderName,
       bool enableMobileNumberLogin = true,
       bool isTrialLicenceKey = true,
-      bool chatHistoryEnable = false,
+      bool chatHistoryEnable = true,
       // int? maximumRecentChatPin,
       // GroupConfig? groupConfig,
       // String? ivKey,
@@ -254,8 +254,7 @@ class Mirrorfly {
     return FlyChatFlutterPlatform.instance.clearAllConversation();
   }
 
-  ///Used as a [updateFcmToken] class for [Mirrorfly]
-  ///used to update FCM Token
+
   static Future<bool?> updateFcmToken(String firebasetoken) {
     return FlyChatFlutterPlatform.instance.updateFcmToken(firebasetoken);
   }
@@ -264,9 +263,13 @@ class Mirrorfly {
     return FlyChatFlutterPlatform.instance.isMuted(jid);
   }
 
-  static Future<dynamic> handleReceivedMessage(Map notificationdata) {
+  ///This [handleReceivedMessage] used to get ChatMessage from FCM Notification [Android] Only
+  ///to Show Notification using FCM [remoteMessage.data] as [notificationData]
+  ///for [iOS] Need to add Notification Extension Service
+  ///add this line in your extension service MirrorFlyNotification().handleNotification(notificationRequest: request, contentHandler: contentHandler, containerID: "xxx", licenseKey: "xxxx")
+  static Future<dynamic> handleReceivedMessage(Map notificationData) {
     return FlyChatFlutterPlatform.instance
-        .handleReceivedMessage(notificationdata);
+        .handleReceivedMessage(notificationData);
   }
 
   static Future<dynamic> getLastNUnreadMessages(int messagesCount) {
@@ -379,14 +382,16 @@ class Mirrorfly {
     return FlyChatFlutterPlatform.instance.getUserLastSeenTime(jid);
   }
 
+  @Deprecated('Instead of use refreshAndGetAuthToken')
+  /// This [authToken] is used to get refreshed Auth Token.
   static Future<String?> authToken() {
     return FlyChatFlutterPlatform.instance.authToken();
   }
 
   static Future<dynamic> registerUser(String userIdentifier,
-      {String token = ""}) {
+      {String fcmToken = ""}) {
     return FlyChatFlutterPlatform.instance
-        .registerUser(userIdentifier, token: token);
+        .registerUser(userIdentifier, fcmToken: fcmToken);
   }
 
   static Future<String?> verifyToken(String userName, String token) {
@@ -645,28 +650,30 @@ class Mirrorfly {
       String? messageId,
       double? messageTime,
       bool exclude = true,
+      bool ascendingOrder = true,
       int limit = 25}) {
     return FlyChatFlutterPlatform.instance.initializeMessageList(
         userJid: userJid,
         messageId: messageId,
         messageTime: messageTime,
         exclude: exclude,
+        ascendingOrder: ascendingOrder,
         limit: limit);
   }
 
-  /// This method is used to Fetch initial conversations between you and a single chat user or group.
+  /// This [loadMessages] is used to Fetch initial conversations between you and a single chat user or group.
   /// This method should be called only after the initializeMessageList Method.
   static Future<dynamic> loadMessages() {
     return FlyChatFlutterPlatform.instance.loadMessages();
   }
 
-  /// This method is used to fetch previous set of conversations between you and a single chat user or group.
+  /// This [loadPreviousMessages] is used to fetch previous set of conversations between you and a single chat user or group.
   /// This set contains the limit/length set in initializeMessageList method
   static Future<dynamic> loadPreviousMessages() {
     return FlyChatFlutterPlatform.instance.loadPreviousMessages();
   }
 
-  /// This method is used to fetch next set of conversations between you and a single chat user or group.
+  /// This [loadNextMessages] is used to fetch next set of conversations between you and a single chat user or group.
   /// This set contains the limit/length set in initializeMessageList method
   static Future<dynamic> loadNextMessages() {
     return FlyChatFlutterPlatform.instance.loadNextMessages();
@@ -721,7 +728,8 @@ class Mirrorfly {
     return FlyChatFlutterPlatform.instance.removeGroupProfileImage(jid);
   }
 
-  static Future<bool?> refreshAndGetAuthToken() {
+  /// This [refreshAndGetAuthToken] is used to get refreshed Auth Token.
+  static Future<String?> refreshAndGetAuthToken() {
     return FlyChatFlutterPlatform.instance.refreshAndGetAuthToken();
   }
 
@@ -1182,6 +1190,11 @@ class Mirrorfly {
   ///if you not mentioned userJid then current user video mute status will be returned
   static Future<bool?> isUserVideoMuted([String? userJid]) async {
     return FlyChatFlutterPlatform.instance.isUserVideoMuted(userJid);
+  }
+  ///Used as a [openAudioPicker] class for [Mirrorfly]
+  ///used to get Audio File in Platform Android Only
+  static Future<String?> openAudioFilePicker() async {
+    return FlyChatFlutterPlatform.instance.openAudioFilePicker();
   }
 
 }