+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [0.0.11] Push Notification
+* handleReceivedMessage method added to get chat message from FCM Notification for Android Only
+* for iOS Need to add Notification Extension Service
+* add this line in your extension service MirrorFlyNotification().handleNotification(notificationRequest: request, contentHandler: contentHandler, containerID: "xxx", licenseKey: "xxxx")
+* authToken Deprecated instead of use refreshAndGetAuthToken
+
 ## [0.0.10] Code Optimization
 * Minified Enabled True Support Added
 * Release Build Bug Fix